% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{anura.data}
\alias{anura.data}
\title{anura.data}
\format{A data frame with five variables:
\describe{
\item{\code{fam}}{Family}
\item{\code{sp}}{The species binomial name}
\item{\code{environment}}{Reproductive environment (still or flowing)}
\item{\code{DF}}{Dominant frequency (hertz)}
\item{\code{SVL}}{snout-vent length (mm)}
}}
\source{
Data downloaded from \url{https://github.com/paternogbc/2015_Rohr_et_al_JAEcol}
}
\usage{
anura.data
}
\description{
A dataset for 509 anura species.
}
\references{
Rohr, D.L., Paterno, G.B., Camurugi, F., Junca, F.A. & Garda, A.A. Background
 noise as a selective pressure: stream-breeding anurans call at higher frequencies.
 Organisms Diversity & Evolution 16, 269-273 (2015).
}
\keyword{datasets}
