% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{anura.tree}
\alias{anura.tree}
\title{anura.tree}
\format{A "phylo" object with 509 tips.}
\source{
Pruned phylogeny downloaded from \url{https://github.com/paternogbc/2015_Rohr_et_al_JAEcol}
}
\usage{
anura.tree
}
\description{
A phylogeny of 509 anura species as a "phylo" object with species
matching "anura.data". Tree pruned from the Amphibia phylogeny
proposed by Pyron and Wiens (2011).
}
\references{
Pyron, R. A.,& Wiens, J. J. (2011).Alarge-scale phylogeny of Amphibia including
over 2800 species, and a revised classification of extant frogs, salamanders,
 and caecilians. Molecular Phylogenetics and Evolution,  61(2), 543-593.
}
\keyword{datasets}
